requirements.txt
fusepy
redis
google-cloud-storage